⚡Frontend: moving Auth bits to differentiate between common and app-specific
