+++ src/Directory/Directory/ClientApp/src/apps/spa/auth/ApiCallTest.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import authorizeService from "services/authorize-service";
+import authorizeService from "auth/service";
 
 /**
  * TODO: PoC only; REMOVE
