+++ src/Directory/Directory/ClientApp/src/auth/Context.js
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+const AuthContext = createContext({});
+
+export const AuthProvider = AuthContext.Provider;
+export default AuthContext;
