+++ src/Directory/Directory/ClientApp/src/components/AuthorizeRoute.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from "react";
 import { useAsync } from "react-async";
 import { Route, Redirect } from "react-router-dom";
-import authorizeService from "services/authorize-service";
-import { Paths, QueryParams } from "constants/oidc";
+import authorizeService from "auth/service";
+import { QueryParams } from "constants/oidc";
 
 const AuthorizeRoute = ({ component: Component, ...rest }) => {
   const { data, isPending, reload } = useAsync(
@@ -16,9 +16,9 @@ const AuthorizeRoute = ({ component: Component, ...rest }) => {
     return () => authorizeService.unsubscribe(subId);
   }, [reload]);
 
-  const redirectUrl = `${Paths.Login}?${QueryParams.ReturnUrl}=${encodeURI(
-    window.location.href
-  )}`;
+  const redirectUrl = `${authorizeService.unauthorised_uri}?${
+    QueryParams.ReturnUrl
+  }=${encodeURI(window.location.href)}`;
 
   if (isPending) return <div>Loading...</div>; // TODO: make sexier
