+++ src/Directory/Directory/ClientApp/src/constants/oidc.js
@@ -1,5 +1,3 @@
-export const ApplicationName = "UKCRC Tissue Directory";
-
 export const Results = {
   Redirect: "redirect",
   Success: "success",
@@ -8,16 +6,4 @@ export const Results = {
 
 export const QueryParams = {
   ReturnUrl: "returnUrl"
-};
-
-const local = "/auth";
-const idp = "/Account";
-export const Paths = {
-  Prefix: local,
-  Login: `${local}/login`,
-  LoginCallback: `${local}/signin-oidc`,
-  Logout: `${local}/logout`,
-  LogoutCallback: `${local}/signout-callback-oidc`,
-  IdpRegister: `${idp}/Register`,
-  Origin: window.location.origin
 };
